Move graph related drawables to drawable.graph
