+++ src/main/java/nl/tudelft/context/drawable/graph/DefaultLabel.java
@@ -1,4 +1,4 @@
-package nl.tudelft.context.drawable;
+package nl.tudelft.context.drawable.graph;
 
 import javafx.scene.layout.VBox;
