+++ src/main/java/nl/tudelft/context/drawable/graph/DrawableGraphNodeLabel.java
@@ -1,4 +1,4 @@
-package nl.tudelft.context.drawable;
+package nl.tudelft.context.drawable.graph;
 
 import javafx.scene.control.Label;
 import nl.tudelft.context.controller.AbstractGraphController;
