+++ src/test/java/nl/tudelft/context/effects/ZoomTest.java
@@ -3,8 +3,8 @@ package nl.tudelft.context.effects;
 import de.saxsys.javafx.test.JfxRunner;
 import javafx.scene.Group;
 import javafx.scene.control.ScrollPane;
-import nl.tudelft.context.drawable.DefaultLabel;
-import nl.tudelft.context.drawable.DrawableNodeLabel;
+import nl.tudelft.context.drawable.graph.DefaultLabel;
+import nl.tudelft.context.drawable.graph.DrawableNodeLabel;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
