+++ src/test/java/nl/tudelft/context/model/graph/GraphNodeTest.java
@@ -3,10 +3,10 @@ package nl.tudelft.context.model.graph;
 import de.saxsys.javafx.test.JfxRunner;
 import nl.tudelft.context.controller.AbstractGraphController;
 import nl.tudelft.context.controller.MainController;
-import nl.tudelft.context.drawable.AbstractDrawableNode;
-import nl.tudelft.context.drawable.DefaultLabel;
-import nl.tudelft.context.drawable.DrawableGraphNode;
-import nl.tudelft.context.drawable.DrawableGraphNodeLabel;
+import nl.tudelft.context.drawable.graph.AbstractDrawableNode;
+import nl.tudelft.context.drawable.graph.DefaultLabel;
+import nl.tudelft.context.drawable.graph.DrawableGraphNode;
+import nl.tudelft.context.drawable.graph.DrawableGraphNodeLabel;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
